use bigint in int_or

+++ macros/udfs/f_int_or.sql
@@ -1,7 +1,7 @@
 {% macro create_f_int_or_null() %}
 CREATE OR REPLACE FUNCTION {{target.schema}}.int_or_null(v_input text)
-RETURNS INTEGER AS $$
-DECLARE v_int_value INTEGER DEFAULT NULL;
+RETURNS BIGINT AS $$
+DECLARE v_int_value BIGINT DEFAULT NULL;
 BEGIN
     BEGIN
         v_int_value := v_input::INTEGER;
